Add to controllers method get one

+++ src/main/java/com/epam/repository/RoomTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoomTypeRepository extends JpaRepository<RoomType,Integer> {
-    Optional<RoomType> findById(Integer id);
+public interface RoomTypeRepository extends JpaRepository<RoomType,Long> {
+    Optional<RoomType> findById(Long id);
 }
